requirements.txt
gradio
transformers
torchvision
requests
google-generativeai>=0.3.0
